+++ lib/tests/processing.rs
@@ -17,7 +17,6 @@ extern crate gong;
 mod common;
 
 use gong::analysis::*;
-use gong::options::*;
 use common::{get_base, Actual, Expected, check_result};
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -32,12 +31,12 @@ fn arg_list_owned_set() {
     // Test works (compiles) using a `String` based slice (as given from `env::args()` for real args)
     // Note, **deliberately** not using the `arg_list` macro here!
     let args: Vec<String> = vec![ String::from("--foo"), String::from("--bah") ];
-    let _ = get_base().process(&args);
+    let _ = get_base().process(&args, None);
 
     // Test works (compiles) using a `&str` based slice
     // Note, **deliberately** not using the `arg_list` macro here!
     let args: Vec<&str> = vec![ "--foo", "--bah" ];
-    let _ = get_base().process(&args);
+    let _ = get_base().process(&args, None);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -88,7 +87,7 @@ fn basic() {
             expected_item!(13, NonOption, "jkl"),
         ]
     );
-    check_result(&Actual(get_base().process(&args)), &expected);
+    check_result(&Actual(get_base().process(&args, None)), &expected);
 }
 
 /// Enforce that option matching is case sensitive
@@ -103,7 +102,7 @@ fn case_sensitivity() {
             expected_item!(1, UnknownShort, 'O'),
         ]
     );
-    check_result(&Actual(get_base().process(&args)), &expected);
+    check_result(&Actual(get_base().process(&args, None)), &expected);
 }
 
 /// Test that everything after an early terminator is taken to be a non-option, including any
@@ -140,7 +139,7 @@ fn early_term() {
             expected_item!(14, NonOption, "-b"),
         ]
     );
-    check_result(&Actual(get_base().process(&args)), &expected);
+    check_result(&Actual(get_base().process(&args, None)), &expected);
 }
 
 /// Test empty long option names with data param (-- on it’s own is obviously picked up as early
@@ -156,7 +155,7 @@ fn long_no_name() {
             expected_item!(1, LongWithNoName),
         ]
     );
-    check_result(&Actual(get_base().process(&args)), &expected);
+    check_result(&Actual(get_base().process(&args, None)), &expected);
 }
 
 /// Test repetition - each instance should exist in the results in its own right. Note, data arg
@@ -180,7 +179,7 @@ fn repetition() {
             expected_item!(7, Short, 'h'),
         ]
     );
-    check_result(&Actual(get_base().process(&args)), &expected);
+    check_result(&Actual(get_base().process(&args, None)), &expected);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -207,7 +206,7 @@ mod utf8 {
                 expected_item!(4, Short, '❤'), // '\u{2764}' black heart
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Some utf8 multi-byte char handling - chars with combinator chars (e.g. accent)
@@ -233,7 +232,7 @@ mod utf8 {
                 expected_item!(7, Long, "ábc"),              // with combinator
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Some utf8 multi-byte char width handling - chars with variation selector
@@ -254,7 +253,7 @@ mod utf8 {
                 expected_item!(2, UnknownLong, "❤️"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Some utf8 multi-byte char width handling - lone combinator chars
@@ -271,7 +270,7 @@ mod utf8 {
                 expected_item!(3, Short, '\u{030a}'),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 }
 
@@ -297,7 +296,7 @@ mod abbreviations {
                 expected_item!(1, AmbiguousLong, "fo"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test handling of abbreviated long options, without ambiguity
@@ -319,7 +318,7 @@ mod abbreviations {
                 expected_item!(3, Long, "foobar"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test handling when abbreviated matching is disabled
@@ -338,9 +337,9 @@ mod abbreviations {
                 expected_item!(5, Long, "foobar"),
             ]
         );
-        let mut opts = get_base().clone();
-        opts.set_allow_abbreviations(false);
-        check_result(&Actual(opts.process(&args)), &expected);
+        let mut settings = Settings::default();
+        settings.set_allow_abbreviations(false);
+        check_result(&Actual(get_base().process(&args, Some(&settings))), &expected);
     }
 
     /// Test that an exact match overrides ambiguity
@@ -369,7 +368,7 @@ mod abbreviations {
             ],
             []
         );
-        check_result(&Actual(opts.process(&args)), &expected);
+        check_result(&Actual(opts.process(&args, None)), &expected);
     }
 }
 
@@ -399,7 +398,7 @@ mod data {
                 expected_item!(4, Long, "help"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test calculation of whether or not short-opt taking data is the last character in the short
@@ -419,7 +418,7 @@ mod data {
                 expected_item!(1, NonOption, "g"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test option with expected data arg, provided in next argument for short options
@@ -447,7 +446,7 @@ mod data {
                 expected_item!(6, ShortWithData, 'o', "def", DataLocation::NextArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test option with expected data arg, provided in same argument for short options
@@ -490,7 +489,7 @@ mod data {
                 expected_item!(9, ShortWithData, 'o', "abx", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test missing argument data for long option
@@ -504,7 +503,7 @@ mod data {
                 expected_item!(0, LongMissingData, "hah"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test missing argument data for short option
@@ -521,7 +520,7 @@ mod data {
                 expected_item!(0, ShortMissingData, 'o'),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test some misc. data handling.
@@ -560,7 +559,7 @@ mod data {
                 expected_item!(8, ShortWithData, 'o', "=b", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test repetition - each instance should exist in the results in its own right. Note, basic
@@ -578,7 +577,7 @@ mod data {
                 expected_item!(4, ShortWithData, 'o', "bc", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test option with expected data arg, declared to be in same argument, but empty
@@ -600,7 +599,7 @@ mod data {
                 expected_item!(3, NonOption, "help"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test option with expected data arg, with data containing `=`. An `=` in a long option arg
@@ -634,7 +633,7 @@ mod data {
                 expected_item!(10, ShortWithData, 'o', "===o", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test argument data that looks like options
@@ -673,7 +672,7 @@ mod data {
                 expected_item!(22, ShortWithData, 'o', "--blah", DataLocation::NextArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test argument data that looks like early terminator
@@ -695,7 +694,7 @@ mod data {
                 expected_item!(5, ShortWithData, 'o', "--", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test long option involving multi-byte chars, to ensure "in-arg" component splitting for
@@ -714,7 +713,7 @@ mod data {
                 expected_item!(5, LongMissingData, "ƒƒ"),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test short options involving multi-byte chars to check offset calculations in iterating
@@ -756,7 +755,7 @@ mod data {
                 expected_item!(13, ShortWithData, 'Ɛ', "b❤", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 
     /// Test the effect of Utf-8 combinator characters - does this break char iteration or byte
@@ -801,7 +800,7 @@ mod data {
                 expected_item!(6, ShortWithData, 'Ɛ', "\u{030a}b❤\u{fe0f}", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(get_base().process(&args)), &expected);
+        check_result(&Actual(get_base().process(&args, None)), &expected);
     }
 }
 
@@ -859,9 +858,9 @@ mod alt_mode {
                 expected_item!(16, NonOption, "-help"),
             ]
         );
-        let mut opts = get_base().clone();
-        opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(opts.process(&args)), &expected);
+        let mut settings = Settings::default();
+        settings.set_mode(OptionsMode::Alternate);
+        check_result(&Actual(get_base().process(&args, Some(&settings))), &expected);
     }
 
     /// Check unexpected and missing data
@@ -881,9 +880,9 @@ mod alt_mode {
                 expected_item!(2, LongMissingData, "hah"),
             ]
         );
-        let mut opts = get_base().clone();
-        opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(opts.process(&args)), &expected);
+        let mut settings = Settings::default();
+        settings.set_mode(OptionsMode::Alternate);
+        check_result(&Actual(get_base().process(&args, Some(&settings))), &expected);
     }
 
     /// Test argument data that looks like early terminator
@@ -901,8 +900,8 @@ mod alt_mode {
                 expected_item!(1, LongWithData, "hah", "--", DataLocation::NextArg),
             ]
         );
-        let mut opts = get_base().clone();
-        opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(opts.process(&args)), &expected);
+        let mut settings = Settings::default();
+        settings.set_mode(OptionsMode::Alternate);
+        check_result(&Actual(get_base().process(&args, Some(&settings))), &expected);
     }
 }